passcode のapiのルーティンングを変えた
